keep searching even after found the score>0.8 for a read. (accurate bkp detection)

+++ pipeline.sh
@@ -18,8 +18,10 @@ if [ -f $original_ref.fai ];then
 fi
 # :<<!
 /mnt/d/breakpoints/script/extract_ref $fq1 $fq2 $original_ref $interval_file $6 $7
-cat ${interval_file}_tmp_* >$interval_file
-rm ${interval_file}_tmp_*
+if [ ! -f $interval_file ];then
+  cat ${interval_file}_tmp_* >$interval_file
+  rm ${interval_file}_tmp_*
+fi
 python /mnt/d/breakpoints/script/get_bed_file.py $original_ref $interval_file
 samtools faidx -r ${interval_file}.bed $original_ref > $extracted_ref
 
@@ -33,19 +35,21 @@ echo Time taken to prepare ref is ${take} seconds. > ${sample}.log
 bwa mem -M -t 5 -R "@RG\tID:id\tSM:sample\tLB:lib" $extracted_ref $fq1 $fq2 > $sample.sam
 python /mnt/d/breakpoints/script/extractSplitReads_BwaMem.py -i $sample.sam >$sample.splitters.sam
 samtools view -h -q 20 $sample.sam > $sample.unique.sam
+rm $sample.sam
 
 end=$(date +%s)
 take=$(( end - start ))
 echo Time taken to map reads is ${take} seconds. >> ${sample}.log
 
 python /mnt/d/breakpoints/script/get_raw_bkp.py -u $sample.unique.sam -o $sample.raw.txt
+!
 python /mnt/d/breakpoints/script/accurate_bkp.py -g $original_ref -u $sample.unique.sam \
 -s $sample.splitters.sam -a $sample.raw.txt -o $sample.acc.txt
 
 end=$(date +%s)
 take=$(( end - start ))
 echo Time taken to execute commands is ${take} seconds. >> ${sample}.log
-# !
+
